requirements.txt
streamlit
requests
python-pptx
comtypes
pywin32